requirements.txt
Django==4.2.1
django-extensions==3.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
selenium==4.9.1
load-dotenv==0.1.0
celery==5.3.0
django-cors-headers==4.0.0